Move new path provider to vist
